Fix major bugs after rebuilding

+++ DataFlow/Models/Order.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
         public decimal Summ { get; set; }
-        public ushort CustomerId { get; set; }
+        public int CustomerId { get; set; }
         public Customer Customer { get; set; }
         public string? Comment { get; set; }
         public ICollection<OrderRows> Goods { get; set; }
